+++ src/CacheService.Core/CacheService.Core.csproj
@@ -4,7 +4,7 @@
         <TargetFramework>net9.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <AnalysisLevel>latest-all</AnalysisLevel>
+        <AnalysisLevel>latest-recommended</AnalysisLevel>
         <InvariantGlobalization>true</InvariantGlobalization>
     </PropertyGroup>
