+++ src/CacheService.Core/Config/ISecretVault.cs
@@ -1,9 +1,9 @@
-﻿namespace CacheService.Configuration;
+namespace CacheService.Core.Config;
 
 /// <summary>
 /// A vault containing secrets for the cache.
 /// </summary>
-internal interface ISecretVault
+public interface ISecretVault
 {
     /// <summary>
     /// An information if secret vault is enabled or not.
